requirements.txt
numpy
pandas
opencv-python-headless
onnx
onnxruntime
pillow
streamlit
imageio
imageio[ffmpeg]